Externalizing default logger
